+++ src/main/groovy/org/liquibase/gradle/LiquibasePlugin.groovy
@@ -55,32 +55,37 @@ class LiquibasePlugin implements Plugin<Project> {
 		// Create tasks that don't require a value.
 		[
 				'changelogSync': 'Mark all changes as executed in the database.',
-				'changelogSyncSQL': 'Writes SQL to mark all changes as executed in the database to STDOUT.',
-				'clearChecksums': 'Removes all saved checksums from the database. On next run checksums will be recomputed.  Useful for "MD5Sum Check Failed" errors.',
-				'diff': 'Writes description of differences to standard out.',
-				'diffChangeLog': 'Writes Change Log to update the database to the reference database to standard out',
-				'dropAll': 'Drops all database objects owned by the user. Note that functions, procedures and packages are not dropped (Liquibase limitation)',
-				'futureRollbackSQL': 'Writes SQL to roll back the database to the current state after the changes in the changeslog have been applied.',
-				'generateChangelog': 'Writes Change Log groovy to copy the current state of the database to standard out.',
-				'history': 'lists out all your deploymentIds and all changesets associated with each deploymentId.',
-				'listLocks': 'Lists who currently has locks on the database changelog.',
-				'markNextChangesetRan': 'Mark the next change set as executed in the database.',
-				'markNextChangesetRanSQL': 'Writes SQL to mark the next change set as executed in the database to STDOUT.',
-				'releaseLocks': 'Releases all locks on the database changelog.',
-				'snapshot': 'Writes the current state of the database to standard out',
-				'snapshotReference': 'Writes the current state of the referenceUrl database to standard out',
-				'status': 'Outputs count (list if liquibaseCommandValue is "--verbose") of unrun change sets.',
-				'unexpectedChangeSets': 'Outputs count (list if liquibaseCommandValue is "--verbose") of changesets run in the database that do not exist in the changelog.',
-				'update': 'Updates the database to the current version.',
+				'changelogSyncSQL': 'Output the raw SQL used by Liquibase when running changelogSync.',
+				'clearChecksums': 'Remove all saved checksums from the database. On next run checksums will be recomputed.  Useful for "MD5Sum Check Failed" errors.',
+				'deactivateChangeLog': 'Removes the changelogID from your changelog so it stops sending reports to Liquibase Hub.',
+                'diff': 'Compare two databases and write differences to STDOUT.',
+				'diffChangeLog': 'Compare two databases to produce changesets and write them to a changelog file',
+				'dropAll': 'Drop all database objects owned by the user. Note that functions, procedures and packages are not dropped (Liquibase limitation)',
+				'futureRollbackSQL': 'Generate the raw SQL needed to rollback undeployed changes.',
+				'generateChangelog': 'Generate a Groovu changelog from the current state of the database.',
+				'history': 'List all deployed changesets and their deployment IDs.',
+				'listLocks': 'List the hostname, IP address, and timestamp of the Liquibase lock record.',
+				'markNextChangesetRan': 'Mark the next change you apply as executed in the database.',
+				'markNextChangesetRanSQL': 'Write SQL to mark the next change you apply as executed in the database.',
+				'registerChangeLog': 'Register the changelog with a Liquibase Hub project.',
+				'releaseLocks': 'Remove the Liquibase lock record from the DATABASECHANGELOG table.',
+				'rollbackOneUpdate': 'Rollback one update from the database (Liquibase Pro key required).',
+		        'rollbackOneUpdateSQL': 'Write SQL to rollback one update from the database (Liquibase Pro key required).',
+				'snapshot': 'Capture the current state of the database.',
+				'snapshotReference': 'Capture the current state of the reference database.',
+				'status': 'Generate a list of pending changesets.',
+				'syncHub': 'Synchronize the local DatabaseChangeLog table with Liquibase Hub.',
+				'unexpectedChangeSets': 'Generate a list of changesets that have been executed but are not in the current changelog.',
+				'update': 'Deploy any changes in the changelog file that have not been deployed.',
 				'updateSQL': 'Writes SQL to update the database to the current version to STDOUT.',
-				'updateTestingRollback': 'Updates the database, then rolls back changes before updating again.',
-				'validate': 'Checks the changelog for errors.'
+				'updateTestingRollback': 'Updates database, then rolls back changes before updating again. Useful for testing rollback support.',
+				'validate': 'Validate the changelog for errors.'
 		].each { taskName, taskDescription ->
 			def commandName = taskName
 			if ( project.hasProperty('liquibaseTaskPrefix') ) {
 				taskName = project.liquibaseTaskPrefix + taskName.capitalize()
 			}
-			project.task(taskName, type: LiquibaseTask) {
+			project.tasks.register(taskName, LiquibaseTask) {
 				group = 'Liquibase'
 				description = taskDescription
 				command = commandName
@@ -89,29 +94,34 @@ class LiquibasePlugin implements Plugin<Project> {
 
 		// Create tasks that do require a value.
 		[
-				'calculateCheckSum': 'Calculates and prints a checksum for the <liquibaseCommandValue> changeset with the given id in the format filepath::id::author.',
-				'dbDoc': 'Generates Javadoc-like documentation based on current database and change log to the <liquibaseCommandValue> directory.',
-				'executeSql': 'Executes SQL in the database given in <liquibaseCommandValue> in this format: -PliquibaseCommandValue="--sql=select 1" or -PliquibaseCommandValue="--sqlFile=myfile.sql"',
-				'futureRollbackCountSQL': 'Writes SQL to roll back <liquibaseCommandValue> changes the database after the changes in the changelog have been applied.',
-				'futureRollbackFromTagSQL': 'Writes (to standard out) the SQL to roll back the database to its current state after the changes up to the <liquibaseCommandValue> tag have been	applied',
-				'rollback': 'Rolls back the database to the state it was in when the <liquibaseCommandValue> tag was applied.',
-				'rollbackCount': 'Rolls back the last <liquibaseCommandValue> change sets.',
-				'rollbackCountSQL': 'Writes SQL to roll back the last <liquibaseCommandValue> change sets to STDOUT.',
-				'rollbackSQL': 'Writes SQL to roll back the database to the state it was in when the <liquibaseCommandValue> tag was applied to STDOUT.',
-				'rollbackToDate': 'Rolls back the database to the state it was in at the <liquibaseCommandValue> date/time.',
-				'rollbackToDateSQL': 'Writes SQL to roll back the database to the state it was in at the <liquibaseCommandValue> date/time to STDOUT.',
-				'tag': 'Tags the current database state with <liquibaseCommandValue> for future rollback.',
-				'tagExists': 'Checks whether the tag given in <liquibaseCommandValue> is already existing.',
-				'updateCount': 'Applies the next <liquibaseCommandValue> change sets.',
-				'updateCountSql': 'Writes SQL to apply the next <liquibaseCommandValue> change sets to STDOUT.',
-				'updateToTag': 'Updates the database to the changeSet with the <liquibaseCommandValue> tag',
-				'updateToTagSQL': 'Writes (to standard out) the SQL to update to the changeSet with the <liquibaseCommandValue> tag'
+				'calculateCheckSum': 'Calculate and print a checksum for the <liquibaseCommandValue> changeset.',
+				'changelogSyncToTag': 'Mark all undeployed changesets up to and including the <liquibaseCommandValue> tag as executed in your database.',
+				'changelogSyncToTagSQL': 'Output the raw SQL used by Liquibase when running the changelogSyncToTag command.',
+				'checks': 'Execute the <liquibaseCommandValue> quality check',
+				'dbDoc': 'Generates Javadoc-like documentation for the existing database and changelogs to the <liquibaseCommandValue> directory.',
+				'executeSql': 'Execute a SQL string or file given in <liquibaseCommandValue> in this format: -PliquibaseCommandValue="--sql=select 1" or -PliquibaseCommandValue="--sqlFile=myfile.sql"',
+				'futureRollbackCountSQL': 'Generates SQL to sequentially revert <liquibaseCommandValue> changes the database.',
+				'futureRollbackFromTagSQL': 'Generates SQL to revert future changes up to the <liquibaseCommandValue> tag.',
+				'rollback': 'Rollback changes made to the database since the the <liquibaseCommandValue> tag was applied.',
+				'rollbackCount': 'Rollback the last <liquibaseCommandValue> changes.',
+				'rollbackCountSQL': 'Write SQL to roll back the last <liquibaseCommandValue> changes.',
+				'rollbackOneChangeSet': 'Roll back the specific <liquibaseCommandValue> changeset, without rolling back changesets deployed before or afterwards. (Liquibase Pro key required)',
+				'rollbackOneChangeSetSql': 'Write SQL to roll back the specific <liquibaseCommandValue> changeset, without rolling back changesets deployed before or afterwards. (Liquibase Pro key required)',
+                'rollbackSQL': 'Write SQL to roll back the database to the state it was in when the <liquibaseCommandValue> tag was applied.',
+				'rollbackToDate': 'Rollback changes made to the database since the <liquibaseCommandValue> date/time.',
+				'rollbackToDateSQL': 'Write SQL to rollback changes made to the database since the <liquibaseCommandValue> date/time.',
+				'tag': 'Mark the current database state with the <liquibaseCommandValue>.',
+				'tagExists': 'Verify the existence of the <liquibaseCommandValue> tag.',
+				'updateCount': 'Deploy the next <liquibaseCommandValue> changes from the changelog file.',
+				'updateCountSql': 'Generate the SQL to deploy the next <liquibaseCommandValue> changes from the changelog file..',
+				'updateToTag': 'Deploy changes from the changelog file to the <liquibaseCommandValue> tag.',
+				'updateToTagSQL': 'Generate the SQL to deploy changes from the changelog file to the <liquibaseCommandValue> tag'
 		].each { taskName, taskDescription ->
 			def commandName = taskName
 			if ( project.hasProperty('liquibaseTaskPrefix') ) {
 				taskName = project.liquibaseTaskPrefix + taskName.capitalize()
 			}
-			project.task(taskName, type: LiquibaseTask) {
+			project.tasks.register(taskName, LiquibaseTask) {
 				group = 'Liquibase'
 				description = taskDescription
 				command = commandName